+++ scraper/scholarship_union/__init__.py
@@ -1,0 +1,3 @@
+"""
+Scholarship Union scraper package.
+"""